analytics-service/requirements.txt
boto3==1.34.34
pyathena==3.5.3
pandas==2.2.0
PyJWT==2.8.0
